backend/requirements.txt
streamlit
openai
sentence-transformers
chromadb
scikit-learn
pandas
python-docx
PyMuPDF
pytesseract
Pillow
PyPDF2
